requirements.txt
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
aiostream==0.5.2
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
Brotli==1.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
cryptography==41.0.5
dataclasses-json==0.5.14
datasets==2.14.6
Deprecated==1.2.14
dill==0.3.7
dnspython==2.4.2
duckdb==0.9.1
duckdb-engine==0.9.2
duckduckgo-search==3.9.3
ecdsa==0.18.0
fastapi==0.104.1
filelock==3.13.0
frozenlist==1.4.0
fsspec==2023.10.0
greenlet==3.0.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.18.0
hyperframe==6.0.1
idna==3.4
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.326
langserve==0.0.21
langsmith==0.0.54
llama-index==0.8.55
loguru==0.7.2
lxml==4.9.3
marshmallow==3.20.1
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nest-asyncio==1.5.8
nltk==3.8.1
numpy==1.26.1
openai==0.28.1
packaging==23.2
pandas==2.1.2
passlib==1.7.4
pinecone-client==2.2.4
pyarrow==13.0.0
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.13
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
SQLAlchemy==2.0.22
sse-starlette==1.6.5
starlette==0.27.0
tenacity==8.2.3
tiktoken==0.5.1
tqdm==4.66.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
uvicorn==0.23.2
wrapt==1.15.0
xxhash==3.4.1
yarl==1.9.2
